Only update address if groups['collivery'] isset() not when disabling module

+++ Collivery/Observer/ConfigDataChange.php
@@ -35,18 +35,20 @@ class ConfigDataChange
         \Closure $proceed
     ) {
         $databaseUsername = $subject->getConfigDataValue('carriers/collivery/username');
-        $newUsername = $subject->getDataByKey('groups')['collivery']['fields']['username']['value'];
+        if (isset($subject->getDataByKey('groups')['collivery'])) {
+            $newUsername = $subject->getDataByKey('groups')['collivery']['fields']['username']['value'];
 
-        $databasePassword = $subject->getConfigDataValue('carriers/collivery/password');
-        $newPassword = $subject->getDataByKey('groups')['collivery']['fields']['password']['value'];
+            $databasePassword = $subject->getConfigDataValue('carriers/collivery/password');
+            $newPassword = $subject->getDataByKey('groups')['collivery']['fields']['password']['value'];
 
-        //if password or username changed by admin user delete authentication file
-        if ($databaseUsername !== $newUsername || $databasePassword !== $newPassword) {
-            $filename = 'collivery.auth';
-            $filepath = $this->filesystem->getDirectoryRead(DirectoryList::ROOT)->getAbsolutePath('cache/mds_collivery');
+            //if password or username changed by admin user delete authentication file
+            if ($databaseUsername !== $newUsername || $databasePassword !== $newPassword) {
+                $filename = 'collivery.auth';
+                $filepath = $this->filesystem->getDirectoryRead(DirectoryList::ROOT)->getAbsolutePath('cache/mds_collivery');
 
-            if ($this->file->isExists("$filepath/$filename")) {
-                $this->file->deleteFile("$filepath/$filename");
+                if ($this->file->isExists("$filepath/$filename")) {
+                    $this->file->deleteFile("$filepath/$filename");
+                }
             }
         }
